requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
CacheControl==0.12.11
cachetools==4.2.4
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
cleo==1.0.0a5
colorama==0.4.6
crashtest==0.3.1
cryptography==38.0.4
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.51.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-jazzmin==2.6.0
django-rest-auth==0.9.5
django-unicorn==0.49.2
djangorestframework==3.14.0
dulwich==0.20.50
filelock==3.8.0
html5lib==1.1
idna==3.4
importlib-metadata==5.1.0
iniconfig==1.1.1
jaraco.classes==3.2.3
jsonschema==4.17.1
keyring==23.11.0
lockfile==0.12.2
more-itertools==9.0.0
msgpack==1.0.4
oauthlib==3.2.2
orjson==3.8.2
packaging==21.3
pexpect==4.8.0
pkginfo==1.8.3
platformdirs==2.5.4
pluggy==1.0.0
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.6.0
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
pywin32-ctypes==0.2.0
redis==4.3.5
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
shellingham==1.5.0
shortuuid==1.0.11
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomlkit==0.11.6
tzdata==2022.6
urllib3==1.26.13
virtualenv==20.16.7
webencodings==0.5.1
whitenoise==6.2.0
zipp==3.11.0
